Generate hashes for files after parsing
Summary:
This second diff is the core of this new approach at incremental initialization.

We introduce a new visitor on the AST which removes all position information and skips method bodies. This maps a regular AST.program to one that is position agnostic and contains only declaration information.

We then hash this AST into a decl_hash.

LSP test failing:

Reviewed By: alexchow

Differential Revision: D5972888

fbshipit-source-id: f5df00c2cbfc75424b11e5a52667979bec0642e2