Fix LSP test flakiness with progress
Summary:
The "didChange" is failing because it expects to get a sequence of messages:
   Hack: preparing to check edits
   Hack: checking edits
   Hack: save any file to do a whole-program recheck
   <blank>
   Error "..."

But sometimes on sandcastle it has been getting this sequence of messages:
   Hack: preparing to check edits
   Hack: checking entire project
   <blank>
   Error "..."

In other words, sometimes (non-deterministically) the Watchman process on sandcastle is telling Hack that a file in the test directory has changed on disk, and this is prompting a recheck.

My solution to this flakiness? -- I'll entirely ignore all progress and actionRequired messages.

Reviewed By: oulgen

Differential Revision: D6071256

fbshipit-source-id: 4e1ce05afa171137955406f6c4c49de10683bd5d