Use MethTabEntries in res::Func
Summary:
A trait method can be renamed when its pulled into a class, so I
switched to using the key-value pair from the method table
itself. Nearly all the uses of the FuncInfo either want the php::Func
or the name, so this works just as well.

I'm putting this up separately from the rest of the trait changes
since it was mostly mechanical, and makes the other diffs easier to
review.

Reviewed By: ricklavoie

Differential Revision: D6071998

fbshipit-source-id: 7ef49d21b6c50c0ab268281fc5f0b9f18a34f528