Better method resolution
Summary:
I realized that I needed to rewrite method resolution to make
unflattened traits work properly in the presence of changes to trait
method attributes. This turned out to be pretty much orthogonal to
traits, and improves method resolution in cases like:

  class A { function foo() {} }
  class B extends A { function foo() {} }
  class C extends A {}
  function f(C $c) { return $c->foo() };

Here A::foo is not marked no-override, so $c->foo() was using a
func-family (as it turns out, a func-family with a single entry - but
we didn't optimize for that case). But with per method-slot
attributes, we see that C::foo really is no-override and resolve the
call exactly.

So again, I'm separating this out to keep the traits diff smaller.

Reviewed By: ricklavoie, alexeyt

Differential Revision: D6088148

fbshipit-source-id: ac85d7ad29d78a28d7b9b851f352867efa81cf80