HHVM compatability mode
Summary:
For features Hack knows how to deal with, but codegen for HHVM should produce
an error, because HHVM does not, we introduce the HHVM compatability mode. An
intended side-effect of this is the short-term perf-win we get from bailing out
quickly (instead of continuing to parse/name/codegen a file, we just bail with
an exception the moment we find features known to be in this category).

Reviewed By: ericlippert, hubyrod

Differential Revision: D6077071

fbshipit-source-id: 57101b83699adec17659a521762fdad922f57f4a