Use member_xval accessors for props at statically known offsets.
Summary: We have code that accesses properties whose offset is known statically (all related to Throwables), for both reading and writing. It's been using propVec and getProp, but we want to kill those because they make it hard to tell if the caller intends to read or write the property. Add propXvalAtOffset methods that return member_xvals, and whose intentions are obvious. Convert this code to use them.

Reviewed By: markw65

Differential Revision: D6070198

fbshipit-source-id: 76e641697f084063e0102ec0ed345c27af8f95d6