Make getProp return member_rval, add getPropLval, convert users.
Summary:
Convert internal callers of getProp that need a PropLookup with a TV* to call getPropImpl directly, and make getPropImpl private.
Convert getProp to return member_rval; convert callers (they all checked that the property existed and was accessible, so just make getProp do this for them).
Add getPropLval returning member_lval for the one caller of getProp that isn't in the ObjectData class and needs to write; this will be reused in later diffs when we convert users of o_realProp.

Reviewed By: markw65

Differential Revision: D6082081

fbshipit-source-id: 274c0f318f9681086042cc7f72d10856f55bc435