Don't build up stdClass instances using o_set.
Summary: A bunch of code that returns stdClasses full of stuff would create an empty one, then o_set properties on it. This is inefficient, and uses the o_set API which we want to kill. Convert all the places we do this to use ArrayInit to build up an array of dynamic props and then ObjectData::FromArray to make the stdClass object as the last step. This kills 124 / 261 instances (47.5%) of o_set in our code, and one of those was a macro used >20 times.

Reviewed By: markw65

Differential Revision: D6103207

fbshipit-source-id: 19647564fecb1b2b9743ee81fe1db4af0e9c16e3