Restructure unit emission a little
Summary:
I made a change which accidentally caused an explosion in the size of
the dependency graph, which resulted in it taking around 6 minutes to
destroy the Index (dbgo mode). This meant that we finished writing the
repo before the hhbbc step finished, but then just sat waiting, rather
than continuing to write the text bytecode, and global data.

This allows those steps to run in parallel with the Index destruction,
and also adds more logging so its clear where the time is being spent.

Reviewed By: ricklavoie

Differential Revision: D6101578

fbshipit-source-id: 41368359aa8e86e2976259ea9fddea3de00bddcb