Better dependency tracking
Summary:
When we're analyzing class-at-a-time, there's no point in marking each
method of the class as depending on some Func, since we'll only add
the class once.

This makes the Index aware of what mode we're analyzing in, and
represents dependencies using a new type which is either a Func or a
Class. The result is fewer DepMap entries, and each entry is more
compact (1 pointer vs 3).

Reviewed By: ricklavoie

Differential Revision: D6101867

fbshipit-source-id: d0dac52227c44dff4526047e21b5e36219b10f7e