Make hhbbc work with unflattened traits
Summary:
This adds the code to import interfaces, constants and methods from
traits, and to prevent over-optimizing trait methods that are going to
be imported into classes at runtime.

It doesn't actually flatten the traits - so is likely to be a
regression when hphpc's option to flatten traits is disabled.

Reviewed By: ricklavoie

Differential Revision: D6072156

fbshipit-source-id: c09dbaf4e163d398b098d29b1ed359b0730773c6