Small refactorings in parser
Summary:
Continuing my pursuit of functorizing the lexer and parser. Here I am simplifying the interface to the precedence parser, simplifying the implementation of the type lexer, and adding a couple more methods to the lexer signature that I'm going to need in subsequent diffs.

The plan here is to start by making the parsers all consume lexers that meet a signature, rather than depending on the minimal lexer directly. I'll then functorize them more fully.

 

Reviewed By: oulgen

Differential Revision: D6104892

fbshipit-source-id: 840260f5da19e6ea2ab259a304e5b355c2a3e0c5