Don't crash when vtable slots collide
Summary:
If vtable slots collide, it means hhbbc mis-allocated them. Assuming
its allocation algorithm isn't broken, that means it must have thought
that some of the classes involved couldn't be instantiated.

Rather than asserting, and bringing the server down, raise a fatal
instead.

I initially planned to just enforce hhbbc's decision by not emitting
the class definition, and emitting a Fatal bytecode instead of the
DefClass[Nop]. But that introduces a bunch of possible confusing
errors; eg hoistable child classes that are defined after the bad
class could end up fataling due to a missing base before we got to the
Fatal bytecode; also, we'd need to replicate the runtime error
messages in hhbbc; and finally its possible that only some
instantiations of the class are known to fatal, while some are
not... in that case, we could still end up with the broken vtables,
and would still have to replace the assert with raise_error.

Reviewed By: ricklavoie, alexeyt

Differential Revision: D6114714

fbshipit-source-id: 68cb64cbaf1761ae9e94159df449f2628f1f3f14