Misc Index struct cleanup
Summary:
At one point, FuncInfo was a std::unordered_map::value_type, so it was
constrained to have a first and second member. Now its just an element
in a vector so we can get rid of first, and just add a func to what
used to be second.

While cleaning that up, I noticed that we have several "private"
structs predeclared in index.h, including FuncInfo, which are only
used by res::Func - so make them members of res::Func instead.

Reviewed By: ricklavoie

Differential Revision: D6116978

fbshipit-source-id: 2f7e241def78ba44f444872495597465cd87ecbe