Convert o_set with no context and static property name to setProp.
Summary:
o_set with no context and a property name that is statically known not to start with a null byte is equivalent to setProp (except that setProp supports native properties and o_set doesn't, but that difference won't break anything). Convert cases that meet these criteria from o_set to setProp; this kills 104/137 (about 75%) of the remaining uses of o_set.
Note that some places took char* and turned it into String before turning it into Variant; if the char* is nullptr this results in a KindOfNull Variant and has to be handled manually by the caller when using setProp.

Reviewed By: markw65

Differential Revision: D6113578

fbshipit-source-id: d23f4e620e530e04143a1604e422dfec2998c8d3