explain why client closed
Summary:
In a few bug reports, I've observed that clientLsp has exited.

I assume this is because there was a rebase to a different version in the .hhconfig, but I don't know for sure.

So I've tracked down all the places in this code where we might exit, and I've made sure they write proper logs to the client before exiting.

Reviewed By: arxanas

Differential Revision: D6095305

fbshipit-source-id: f10c81aeef10cd8aa6660c02374585f751087239