show progress for informant delays
Summary:
It used to be that you could attempt a call to `ClientConnect.connect()` and, if the monitor was present (and hence the lockfile present) then you'd get an answer immediately. At least, you'd get an answer from the monitor; you might still need to wait for the server after handoff.

The existing code doesn't do a careful job of handling the case where lockfile is present but no answer is forthcoming from the monitor. It simply shuts down and invites the user to click restart.

But now with informant we have a new situation where the informant will delay doing the monitor handoff until after everything has settled.

This situation is fairly common (it occurs often when a user rebases past a change in hhconfig). I've been getting several bug reports of it a week. So we need to do something better.

This diff adds a progress indicator while we're waiting for the informant to let us connect to the monitor.

Reviewed By: arxanas

Differential Revision: D6106319

fbshipit-source-id: d49ac5e92ca45f2f56ee6c6e91fe0bf0fb575b1a