Fix typeSame when both inputs are known to be static
Summary:
typeSame called loosen_staticness on its inputs so that eg if one was
BCStr, and the other was BSStr, it wouldn't (incorrectly) return false.

This threw away too much information; if one's a known static
string, and the other is an optional known (but different) static
string, we can definitely return false.

This adds loosen_unstaticness, which just adds the static part to
anything with a counted part; we should never generate anything with
only a static part to start with (because later optimizations may be
able to determine an exact value, which would then be static, which
would not be a subtype of the earlier type), but I'm not confident
that the code base is 100% clean on that front - so this seems like
the safest quick fix. I'll follow up by adding assertions to
Type::checkInvariants, cleanup some tests that intentionally create
counted-only types, and deal with any other fallout.

Reviewed By: ricklavoie

Differential Revision: D6125132

fbshipit-source-id: e50ad5cac093afcacc6b4d30b30c16a7891208d2