Refactor jsonrpc part 1: input
Summary:
I want to make "jsonrpc" be a general-purpose re-usable library that handles both input and output of jsonrpc over stdin and stdout.

Currently we have module "jsonrpc_queue" which is used for the input queue of jsonrpc messages: receiving requests, notifications, responses. This module is re-used by Flow.

We also have clientLsp.ml which contains functions for sending jsonrpc messages: sending responses, notifications, and requests. Also for, when we do receive a response, for triggering the appropriate request callback.

I want to re-use those functions for sending jsonrpc inside Flow as well.

This diff is the first of two steps in refactoring. It turns jsonrpc_queue into a functor (since that will be needed for the second step). And that's about it. Just boring rewrite.

Reviewed By: arxanas

Differential Revision: D6006463

fbshipit-source-id: 623afa19273ec74e400ed084093079249e39e698