Refactor jsonrpc part 3: lsp wrappers
Summary:
There are a few LSP wrapper-abstractions that I want to re-use in Flow. So I'm pulling them out into a common library.

The main ones are wrappers for showMessage request, progress notification, and actionRequired notification. They're a bit cumbersome.

(I don't want to write a comprehensive typed wrapper for LSP. That'd be a lot of work for little benefit. Only the bits that benefit from it.)

Reviewed By: arxanas

Differential Revision: D6006469

fbshipit-source-id: a395a1d86d509aaaa764f85b937d514ee882fc97