HHVM Parsing for inout parameters
Summary:
Implements support for inout parameters on functions and methods in the hhvm frontend compiler. In particular the inout annotation must occur at both the callsite and the definition site, cannot occur on async functions, generators, functions with ref parameters, or memoized functions (this restriction is somewhat arbitrary and based mostly on the fact that we don't allow ref parameters on these functions, I will revisit this in a later diff).

These changes solely affect the parser and scanner, and in particular do not add any of the requisite runtime support for this feature. Runtime support can be found in D6002807.

The detailed specification of this new feature can be found here: https://fb.quip.com/3utHAy5kEs8z

Reviewed By: swtaarrs

Differential Revision: D5920433

fbshipit-source-id: fc1684fbcdde908f7e6a4789d811c653b811125f