Update memcached to set MEMC_VAL_COMPRESSION_ZLIB flag
The PHP implementation of memcached cannot get/decompress values that were set/compressed by the HHVM implementation of memcached. This is because the HHVM compressed payload uses the "old-style" format that php-memcache does not understand. This commit changes the compressed payload to use the "new-style" format, which both PHP and HHVM understand.