Stop sharing WatchmanClients between ext_watchman PHP calls
Summary:
In the ext_watchman PHP API there are no explicit functions for managing connections to a Watchman server. Instead the user has to rely on the extension doing the right thing, opening and closing connections as needed. This API is deliberate as some connections need to persist between requests to handle subscriptions.

At the moment ext_watchman tries to aggressively re-use connections. However this code is complicated and has lead to abominations like D4809472. I tried to clean up the code in D4816886 but this just created even worse bugs.

In hindsight trying to reuse connections aggressively is a bit of a premature optimization - the benefits of which are limited as opening and closing Unix sockets is fairly cheap anyway. So, this diff removes the connection sharing feature and just re-opens new sockets to the Watchman server as needed.

Reviewed By: paulbiss

Differential Revision: D6104264

fbshipit-source-id: 38dd0c1861b5a6277f3e6d355a0665055963004c