autocomplete should return "incomplete" in more places
Summary:
Some users had been reporting that autocomplete's typeahead was giving incomplete results. Here's how it works...

1. You type `V` and hh_server gives back a list of top-level classes and functions that start with the letter `V`. It doesn't give back *all* of them; if there are too many then it truncates the list and marks it as an incomplete list.

2. You type further letters to make `Vec`. If the first list had been complete then Nuclide will merely filter it down on the client; if the first list had been incomplete then Nuclide will issue a new autocomplete request to hh_server.

The logic in the server is a bit funky. It first looks for maximum 25 trie prefixes of the answers, finds what files they're in, and returns all top-level classes+functions that are defined in those 25 files. But if more than 1000 classes/functions are defined in those files then it marks its answer as "incomplete".

What was missing: if there were more than 25 trie prefixes of answers, then it should also mark the results as incomplete as well. That's what this diff does.

I wanted to add a test for this. But thought it was too risky because it might be unstable. The test would have 26 classes "Aaa...Aaz" in one file, and 26 classes "Aba...Abz" in another file, and in a third file it would insert the letter "A" and ask for completion. But which trie prefixes would it pick? Would it pick 25 trie prefixes alphabetically, or in order of discovery? Didn't want to risk it.

Reviewed By: arxanas

Differential Revision: D6133377

fbshipit-source-id: 7ea5f2a18596d6aacdb795188fda6661185d2263