one-sided exceptional control flow
Summary:
Allow semdiff to compare functions in which the structure of
the exception handlers on the two sides doesn't match up precisely, so
exceptions are treated more like jumps already are. We use a new special
pc ([],-1) to indicate that one side has already thrown an exception
out to the caller's frame while we're propagating on the other side.

Refactoring of one-side and two-side rules also sets things up for
forthcoming step-indexing changes.

Reviewed By: hubyrod

Differential Revision: D6076713

fbshipit-source-id: 2c3ea722565aa5bb818919de3a9392c7ced161ea