put lax unset behaviour behind a flag
Summary:
Make semdiff be stricter about UnsetL instructions matching up, to account
for potential observability of finalizers. The previous, laxer behaviour now enabled with a
new --laxunset flag.

The scripts for testing the codegen run without the flag, those for semdiff itself
(in run_semdiff_tests) run with the flag.

Reviewed By: siegebell

Differential Revision: D6078153

fbshipit-source-id: 2bb7a95869fe82a7351dad43d21e030ef21dd4ce