Continuing to functorize the parser
Summary:
In this diff, all the parsers are parameterized in the kind of lexer they use and pass around to each other, but the lexers are still constrained to produce minimal tokens.

In upcoming diffs we'll abstract this away further, so that parsers are not just parameterized in lexers, but also in the kind of syntax trees they produce.

Differential Revision: D6107028

fbshipit-source-id: cc650203ae023f41f20a6039c5e193bee925fd8c