Continue functorizing the parser
Summary:
This undoes a mis-step I made in the previous diff. I need to be able to parameterize things in *both* syntax kind and lexer kind, because there can be two lexer types -- the normal lexer and the lexer which handles type lexing -- and this varies independently of what sort of token the lexer is producing.

The correct way to envision things is that a Syntax provides a Token, and a compatible Lexer is any lexer that can produce that token type.

The next step is to add factory methods to the syntax signature; we can then parameterize the parsers to call factory methods via the signature rather than hard-coding them to the minimal syntax factory.

 

Differential Revision: D6128830

fbshipit-source-id: c6e401ef3137c6a2cac102f3e1ae3f1915d51eeb