Dont restart if found saved state in XDB is too distant
Summary:
Prior stack set things up so that XDB lookup could occur inside
the informant. Initially, we used that information to only
decide whether or not to restart based on the existence of a saved state.

Now we make it slightly smarter by requiring the saved state distance
to be closer than the incremental distance.

Reviewed By: jamesjwu

Differential Revision: D6070974

fbshipit-source-id: 7a9ced5dd9f20060bad59d87dba2305e6187c3e2