Create `hh --lint-stdin` to calculate lint warnings for file contents provided on stdin
Summary:
The motivation behind this is to make `hh --lint` into an Aurora lint rule. Aurora is already a generic lint processing engine. In order to simplify the architecture of the Aurora language server for streaming diagnostics, I need to treat all lint rules uniformly, instead of special-casing `hh --lint`.

This change allows us to pass in the contents of the file as they are in the editor, even if they haven't been saved to disk.

Note that this still requires us to invoke `hh --lint-stdin` with a filename, e.g.

```
cat myfilecontents | hh --lint-stdin myfile.php
```

even though we are passing the file contents on stdin. We still need to do this because some lint rules rely on the name of the file (for example, the one that warns that if you declare only one symbol in a file that the file should be name after the symbol).

Reviewed By: ljw1004

Differential Revision: D6107112

fbshipit-source-id: 5fdcfc96c875c54c35033f3777e5e081b9568d0e