Add ignored_paths config option
Summary:
This adds a ignored_paths feature to the .hhconfig: add a list of string regexps. Any file whose absolute path matches the regexp will be ignored by the Hack server.

This actually allows us to remove filesToIgnore from facebook only as a stub by adding the single "phpunit" ignored tests to the list of ignored path regexps instead.

I'll go back and delete filesToIgnore once this diff makes the release, and we change the .hhconfig to include phpunit.

Ignored files *don't exist* to the typechecker.

Reviewed By: alexchow

Differential Revision: D6142929

fbshipit-source-id: a6ebef6b1a4eb6654179c3dca4baaba224126a23