Continue functorizing the parser
Summary:
The parser helper library -- which provides services like "parse the next token and give an error if it is unexpected", and so on -- is now functorized in the kind of syntax trees it produces.

It only produces missing, token, list and list item trees; this is pretty straightforward.

Getting the statement, expression, type and declaration parsers functorized -- finally -- is within striking distance. 
 

Differential Revision: D6164263

fbshipit-source-id: 1c13502665142d6dd4ff15b4bca6d1e9794ea9a0