Introduce AwaitAll bytecode for genva
Summary:
Similar to D6158265, remove the other packed array (which contained the unawaited waithandles). We were constructing this array and passing it to `c_AwaitAllWaitHandle::fromArray()`, which extracted all of the waithandles, checked that they were instances of `c_WaitHandle` and stored them into a newly allocated `c_AwaitAllWaitHandle`. Because of the way we've structured our inlined genva bytecode we already know that all of these cells contain instances of `c_WaitHandle`, furthermore, we don't need the array at all as the values ultimately end up inside the `c_AwaitAllWaitHandle`. Lastly, in the case where all of the waithandles are already finished we still constructed the `c_AwaitAllWaitHandle` which simply produced a static waithandle with the value null.

Originally this code used a CMANY stack flavor for `AwaitAll`, however I observed that the refcount optimizer had trouble eliding IncRef/DecRef pairs on the fast path (no suspend), and worse, couldn't determine that all of the DecRefs were DecRefNZs (because the locals still held references to the waithandles). Instead this version uses a local-range (as the temporary locals were all allocated in sequence by the emitter), and the new `CreateAAWH` IR instruction extracts and IncRefs locals as necessary (this also prevented an extra unnecessary spill of all of the waithandles to the stack).

Reviewed By: jano

Differential Revision: D6159186

fbshipit-source-id: 5c48041c96c318909142fb663474b2502d02b8bc