Add needs_redecl to list of oldified_defs so that we don't oldify files twice
Summary:
After a long dig, I figured out a consistent repro of the sharedMem failures:

1. edit a file that has a dependency on a second file, in a way that only changes declaration positions. This adds the second file to needs_decl.
2. open the second file in the editor, and make a change that changes positional declarations.

Doing the above causes file2 to be oldified twice: once in needs_decl and a second time because the file needs to be re-oldified upon declaration. I was unaware of the invariant that oldified files cannot be oldified again, which led to this error.

P.S. I'm not 100% sure why semantically it should require *two* edits of file 2 to trigger the error on the unit test, but semantically this fixes the bug

Reviewed By: dabek

Differential Revision: D6180804

fbshipit-source-id: 05dcff8d8fb0055c64c3b6a4411cd6fa8647de58