Refactor native state loading in server init. Catch all native load exceptions
Summary:
Refactor native state loading in server init.

Just a small refactor with a tiny behavioral change.
Now catching an exception happens further up the call stack,
which catches an exception while making the State_loader future as well.

Currently, when no state is found, an uncaught exception is thrown which
crashes the server. This change will catch that exception.

Reviewed By: dabek

Differential Revision: D6168229

fbshipit-source-id: b06a02cc19831e68308fd52edcaf4b75e2492ad4