Change `Pos`itional information of pragmas in FFP lowerer for parity with legacy
Summary: Pragmas (`HH_IGNORE_ERROR` / `HH_FIXME` / ...) are treated differently in `Parser_hack` when it comes to positional information; notably, their end-position is the character before the start-position of the first non-trivia afterwards. Also, the beginning-of-line number is off-by-one. This diff makes the lowerer of the Full Fidelity Parser line up with that behaviour.

Reviewed By: ericlippert

Differential Revision: D5913918

fbshipit-source-id: a4951f3d3d9a4d1345461c51d3dc4893fa3b6a8a