Hack type checker: don't use contextual inference for return types if not present
Summary:
Contextual type inference kicks in on `return` statements even when the enclosing lambda doesn't have an explicit return type.
This causes incorrect errors to be reported. For example,
```
<?hh // strict
// Copyright 2004-present Facebook. All Rights Reserved.

function throws<T>(TypeStructure<T> $ts, mixed $value): T { ... }

function testit(mixed $value, TypeStructure<Map<string, int>> $ts): void {
  $f = (int $arg) ==> {
    if ($value === null) {
      return null;
    } else {
      return (throws($ts, $value));
    }
  };
}
```
Here, there should be no error, but the type-checker attempts to use the return type (a `TVar`) of the lambda
to guide the inference of the return expression.

The fix is to add a boolean to the `return` field of the typing environment that indicates whether the type is
to be used for contextual inference or not.

Reviewed By: nbenton

Differential Revision: D6196380

fbshipit-source-id: bea2a0fe19a683416a7d2226cf1e292c4053c25f