Fall over to normal state if tiny state doesn't exist
Summary:
Don't just give up if you can't find a tiny state, try to use the normal one, and log it.

We could probably be smarter about falling over here, and fall over a bit earlier so as to not have to query xdb twice, but this way is simpler and easier, and we don't expect to need to fall over unless we need to turn off sandcastle saving tiny states quickly.

Reviewed By: alexchow

Differential Revision: D6205022

fbshipit-source-id: 93f70b07600684667cebf0294c65122e308fb918