Continue functorizing the parser
Summary:
The parser is functorized! We can now create parsers and get parse trees out given any syntax tree kind that has tokens that can be produced by the lexer.

So we're done, right?

Not yet! That "can be produced by the lexer" bit is now the gating factor. I'll need to update the positional and minimal token modules so that they have a consistent set of factory methods, and that the lexer uses those methods.

 

Differential Revision: D6177349

fbshipit-source-id: 3dd9c79ac793ded363559fe4b6ced922b972dd51