Preclass doesn't need instanceCtor/instanceDtor.
Summary: We had instanceCtor/instanceDtor function pointers on both class and preclass (the preclass ones would be copied onto class when it was created), but the only thing that was using the preclass ones is Closure, and it was always setting them to the same thing (modulo repo mode). Make Closure set them on the class (like Collections and WaitHandles do), and remove instanceCtor/instanceDtor support from preclass.

Reviewed By: markw65

Differential Revision: D6207244

fbshipit-source-id: a7519eaae2b8bd72a51f7d522fb7f09b78786a2c