Fold allocation into ObjectData::clone to avoid redundant work.
Summary: Fold allocation into ObjectData::clone, and cleanly separate it into isCppBuiltin(), HasNativeData and normal object cases. This keeps us from having to do some of the same checks twice (in ::clone and again in ::newInstance), and we can skip the class attrs/class init checks at the start of ::newInstance (since we already have an instance of the class, they already passed before). There is no need for throwable_init when cloning a Throwable (which calling ::newInstance would invoke). Finally, there was no way to get to the code that tries to invoke __clone at the bottom of clone() with a isCppBuiltin(), so stop trying to handle that. It is possible to set dynamic properties on generators https://3v4l.org/3HUsQ so add a unit test for doing so and then cloning the generator.

Reviewed By: markw65

Differential Revision: D6196389

fbshipit-source-id: 02a47c4caaf294f5b3bd4f7fb82049c6e3971dbd