Revert changes to serverTypecheck
Summary:
Considering that we aren't using incremental init and hashing works better, the changes I made for this behavior aren't really worth the extra complexity, or the crashes, so I'm reverting it for now.

We can replace it with a check that uses the file hashes that we already made from initialization, if we'd like.

Reviewed By: ericlippert

Differential Revision: D6209734

fbshipit-source-id: 0ef347edabccdb8b11b62617a65883eda877d0f7