Continue functorizing the parser
Summary:
The positioned token and trivia modules need to be brought up to support the functions necessary to be constructed directly in the lexer.

While in there, I noticed that a method in the signature was only called in one place, and could be simply moved into the module that calls it.

Differential Revision: D6178547

fbshipit-source-id: d84d7aab21a5b28602c5cd32925794eac9beb725