fails to read hhconfig version if server absent at startup
Summary:
The hh_client lsp process is meant to read hhconfig's version field at startup. The idea is this:

(1) we launch "hh_client lsp" from the directory containing .hhconfig. This allows the find_hh.sh script to launch the correct version of hh_client according to the version that's written in the .hhconfig file.

But at this stage the hh_client process doesn't yet know its root or its version.

(2) Nuclide sends "initialize" LSP request. At this stage, hh_client knows its root, and reads the version from the .hhconfig file in that root, and connects to the hh_server for that root too.

(3) At any stage in the future, if hh_client detects that hh_server has shut down, then it re-reads the version from .hhconfig. If the newly-read version differs from what it read at startup then hh_config terminates, trusting Nuclide to relaunch it.

That's the intent. The code for it was messed up before. It used to read hhconfig version only as part of a succesful attempt to connect to hh_server; therefore, if hh_server was absent or blocked due to informant then it simply didn't read hhconfig version at all.

I don't know why it was written that way. It was bad. In cases where hh_server was absent or blocked due to informant, then hh_client would just terminate itself immediately.

This diff fixes the problem with a cleaner structure for when to read hhconfig version.

Reviewed By: arxanas

Differential Revision: D6211683

fbshipit-source-id: 4ef4b574c3b81a5c494425327d9cf24d1eb0b72e