Add Hg.rev type; use everywhere
Summary:
This adds a Hg.rev union type around hg_rev and svn_rev, and
uses it throughout.

In a followup, we would like to load saved state canaries, based off
a hg_rev hash instead of finding the state for the nearest svn_rev.
So we need this first.

Reviewed By: jamesjwu

Differential Revision: D6179379

fbshipit-source-id: f30442ff3567d25cc7d20f5bbfc3484e4fe375a2