Continue functorizing the parser
Summary:
The lexer, token signature, positioned token and minimal token modules now all agree upon what is needed to build a token.

In order to build a positioned token we're going to need the source text and offset, so those are now passed in to the token factory, and ignored in the minimal token.

There are a couple of places with TODOs where the token is built post-lexing by the parser; those will need to be updated to fetch the offset and source of the original token.

Differential Revision: D6179179

fbshipit-source-id: e9baafcb3adea9d534ec564346b7d233caba0035