Convert more o_realProp to getProp.
Summary:
Calling o_realProp with the ObjectData::RealPropUnchecked flag does not make sense when you're getting a declared property of a class in the context of that class: callers were disabling the very visibility/accessibility check that providing the context makes them pass. Since they don't need to pass this flag, these callers can be converted to getProp.
Also add some assertx in ext_spl that jan asked for on an earlier diff.

Reviewed By: mxw

Differential Revision: D6169297

fbshipit-source-id: c3767226eefcb3163151167afbd29a3ac679291b