Don't break heap scan when cloning generators.
Summary:
jan correctly pointed out that on D6196390 I forgot to properly initialize the m_aux32 field in the NativeNode of the cloned generator. This value is used when scanning the heap to find the ActRec for the generator, and if it's left as 0 the scan won't find the Generator's locals, so they may be collected prematurely. Make Generator::operator= set this value correctly, and add a unit test to test generator cloning interacting with GC.
NativeNode was referring to the field that stores the offset to the ActRec as m_count, but NativeNode is not refcounted. m_count overlaps with m_aux32, which is the name we should use in non-refcounted heap objects; make it so.
As jan pointed out, nativeDataInstanceCopyCtor is getting the total size of the native data for the Generator in an inefficient way; make it use ndsize.
ndextra was dead code, remove it.
Generator::Create was needlessly templatized on bool clone; nothing ever used a value other than false, even before my earlier refactors. Remove useless templatization.
Finally, add a unit test that verifies that cloning async generators fatals. This would segfault before D6196390 because we would partially construct the clone before throwing in Native::nativeDataInstanceCopy, and then blow up while trying to destroy the partly constructed clone.

Reviewed By: jano

Differential Revision: D6219696

fbshipit-source-id: b21cc7ef4ecb6ac4ac7661eef49d48a16880addb